RAG/requirements.txt
langchain==0.3.13
pypdf2==3.0.1
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
numpy>=1.26.2
torch>=1.11.0
transformers>=4.41.0
flask==2.0.1
werkzeug==2.0.3
requests==2.31.0
python-dotenv==1.0.0
